+++ flutter_ui_bloc/lib/flutter_ui_bloc.dart
@@ -3,7 +3,6 @@ library flutter_ui_bloc;
 export 'package:cross_file/cross_file.dart';
 export 'package:flutter_form_bloc/flutter_form_bloc.dart';
 
-export 'src/form/SubmitButtonFormBuilder.dart';
 export 'src/form/fields/base/ChipGroupFieldBlocBuilder.dart';
 export 'src/form/fields/base/CircularCheckboxFieldBlocBuilder.dart';
 export 'src/form/fields/base/DurationFieldBlocBuilder.dart';
@@ -11,4 +10,5 @@ export 'src/form/fields/base/InputFileFieldBlocBuilder.dart';
 export 'src/form/fields/base/SliderFieldBlocBuilder.dart';
 export 'src/form/fields/base/upload_file_field_bloc_builder.dart';
 export 'src/form/fields/single/tab_bar_field_bloc_builders.dart';
+export 'src/form/submit_button_form_bloc_builders.dart';
 export 'src/form/typedefs.dart';